requirements.txt
# Core
pydantic>=2.0.0
pyyaml>=6.0

# Data sources
feedparser>=6.0.0
httpx>=0.25.0

# Browser automation (for PDF download)
playwright>=1.40.0

# PDF processing
# marker-pdf  # Install separately due to ML dependencies

# LLM APIs
openai>=1.0.0  # For DeepSeek-compatible API
zai-sdk>=0.1.0  # Zhipu official SDK

# Utilities
python-dotenv>=1.0.0

# Step 4 hybrid filter (local vectors)
numpy>=1.26
sentence-transformers>=3.0.0
torch>=2.5.1  # Install the matching CUDA wheel manually for GPU acceleration (see PyTorch download instructions)

# Testing
pytest>=9.0.0
